Move HTML chunking to the html-utils module
The chunking logic is improved:

1. A first pass uses a "chunking" mode to return a sequence
of document nodes. By default, a chunk is created for each
except the bodymatter) and each sub-section
of the bodymatter.

The chunking is therefore based on templates rather than a single
XPath, which allows to 'normalize' mixed layers (e.g. non-section
siblings of top-level chunked sections are grouped in a section
to be part of their own chunk.

2. A second pass creates the result documents. The IDREFs are fixed
by looking-up the chunk document nodes with an XSLT key.
The output document names are based on the source name, with a number
suffix corresponding to the chunk position in the seqeuence of chunks.