Improved figure translation.
The rules are:

- captionable elements: objects or tables
- captioning elements: hd, caption, citation

A figure is created for each block with a 'figure' role
or for each captionable element with associated captioning
content.

Within a figure containing mulitple captionable elements,
if some of them have dedicated captioning content (i.e.
captions associated to them only and not the other objects
in the figure), then a sub-figure children is created.
This matches the practices shown in the examples of the
HTML5 specification.

Fixes Issue 234.